+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - PR #5974 Use libcudf instead of cupy for `arange` or column creation from a scalar.
 - PR #5874 Add `COLLECT` groupby aggregation
 - PR #6119 Add support for `dayofweek` property in `DateTimeIndex` and `DatetimeProperties`
+- PR #6125 Add support for `Series.mode` and `DataFrame.mode`
 
 ## Improvements